requirements.txt
Flask==2.0.1
Flask-QRcode==3.2.0
qrcode==7.3.1
fpdf2==2.4.5
Pillow==9.0.0
pycryptodome==3.12.0